sw: fix assert on layout of ooo63564-1.odt
The problem is that some SwTextFrames (1415 in particular) inside the
follow-flow-row of a table move backwards, during(!) which the
follow-flow-row is deleted; i.e., the SwTextFrame was already moved
backwards by the pNewUpper->Calc() in MoveBwd(), then it is Cut and
Pasted at the end of pNewUpper, when it is actually already a lower,
somewhere in the middle...

This triggers the assert in SwFrame::PrepareMake() that was added in
commit 5f3bf6aea077be0a79569dba42df4ee052950608

Prevent this by using the IsDeleteForbidden() check that was introduced
in commit 9976934286c883b162fca32fb42967f2b0aa245f; this is similar to
how MoveBwd() already locks section frames to prevent such problems.

Ultimately a regression from 608fc3f8616a704f68d120ea41f5d95b8316b3be.

Change-Id: I893ec3e491b4cbe1569edf97fec31d3dd74548ed
Reviewed-on: https://gerrit.libreoffice.org/74298
Tested-by: Jenkins
Reviewed-by: Michael Stahl <Michael.Stahl@cib.de>