sw comments on frames: fix comment insert for as-char frame at para start
This adapts SwWrtShell::InsertPostIt() to behave similar to commit
7b2312d4bcd09e488c5e53b9531dd1439f691645 (sw comments on frames: delete
comment of frame when deleting frame, 2019-07-03), i.e. instead of
hoping that the cursor will be at the end of paragraph and traveling
back, just set the cursor to the remembered anchor position.

This is cleaner, and as a side-effect also fixes the scenario when
creating a comment on an as-char image, which happens to be at the start
of the paragraph.

Change-Id: Iedca0feb62242677b6e8b69ef7b813d6da72c8eb
Reviewed-on: https://gerrit.libreoffice.org/75093
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins