WIN get embedded font info from EnumFontFamiliesEx
Instead of some manual decoding of the .fot file, this extracts
all the relevant information of the registered font using the
default routines already used when enumerating the system fonts.

We still have to extract the font family from the .fot file to
know the registered family name of the embedded font.

And it uses the OSL functions to create the temporary font resource
file name and generally adds more SAL_WARNs when some part of the
private font registration fails.

Change-Id: I5e118bf1f3bc1fb7bbf4c68a383a45892dffaf98
Reviewed-on: https://gerrit.libreoffice.org/75102
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>