qtcreator: Make paths work with cygwin
Absolute paths in files created by GbuildToJson
in Cygwin environment are Windows paths
(like "C:/....'), while relative paths in the Python
program executed in Cygwin are treated as Unix paths
('/cygdrive/c/....') which caused wrong relative paths to
be generated by the call to 'os.relpath'. It would walk up to
the root of the Windows path and then walk down the tree in the
Unix file system path again.

This converts the paths to absolute Windows paths first to avoid
this issue.

Change-Id: I2e3e6926e312d64aa18067933a5903ac7ad5d31a
Reviewed-on: https://gerrit.libreoffice.org/75114
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>