sw lok: hide UI to set Word-incompatible anchor types
Let's make it a bit harder for users to hurt themselves, at least in
Online. Affects the context/popup menu and the dialog of frames.

This is similar to how certain border types are hidden on the UI,
depending on if MSO supports them. Though that goes a bit further and
disables those types even on the desktop.

Change-Id: I6b9205ef3df8a7bc15fbcf787d134743c09e98da
Reviewed-on: https://gerrit.libreoffice.org/75129
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins