tdf#126255 disable broken relative link unit test
Relative links are currently broken and this was hidden by this
unit test. Funnily asserting for aURL.startsWith("..") happens to
work for most tested build targets, but the link is generally
wrong, even if it starts with "..".

I don't know either how to retrieve the relative link from the XML
nor how to generate the correct link in Calc, so this unit test
can test, if the linked document exists and work correctly.

So - hopefully as a temporarly measure - this disables the
comparing assertion, until the bug and test are fixed.

Change-Id: I165ac3ce763133fcef6cea5a8dd221b0076f1ee4
Reviewed-on: https://gerrit.libreoffice.org/75172
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>