tdf#126234 up relative bullet size limit to 400%
The legitimate value for MS Office document bullet relative size is between 0% and 400%
But for ODF the value is normally between 0% and 250%.
Per ODF 1.2 spec "19.761 text:bullet-relative-size" there is no limit placed on the
relative size of the bullet. This change is to make the upper limit match MS Office
to address the import issue.

UI also changed to reflect the raise

Change-Id: Ic4bbef47e86a04b9a6f9f6188326c8b66e034807
Reviewed-on: https://gerrit.libreoffice.org/75189
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>