tdf#126173 RTF import: fix lost SHAPE fields
Commit 94a924aaea960ad94da72ef5bbd47ba7bec92e79 (fdo#82860 RTF import:
fix handling of SHAPE fields, 2014-10-15) already tried to handle this,
but aCode is the shape command + its parameters (SHAPE  \* MERGEFORMAT)
for the bugdoc, while what we want is just the shape command.

The field variable already contains a tokenized version, which was used
previously only to decide if a field is unhandled or not.

Reuse that for the shape comparison, so bugdoc's shape with parameters
also appears.

Change-Id: I7e044b94bcfab490c956b33c11dd6c69443939f5
Reviewed-on: https://gerrit.libreoffice.org/75243
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>