Qt5 correctly draw rectangle frame
This bug can be seen with the Writer comments "button" to hide
or show comments, where qt5 misses the bottom and right borders.

The drawRect from X11SalGraphicsImpl and SvpSalGraphics do this
and it fixes the visual glitch. I'm not sure this is needed for
drawAlphaRect, as other implementations don't handle drawing
borders in these at all.

Change-Id: Ic82d499cd8e1bb420a7df14269e86c75e0d30c0b
Reviewed-on: https://gerrit.libreoffice.org/75416
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>