qtcreator: Recursively include module's header files
Recursivley walk the include directories located
inside the current module's directory, since includes
are basically paths relative to the include directories
and can refer to files in subdirectories of the include
path, like e.g.

    #include <extended/AccessibleBrowseBoxBase.hxx>

in 'accessibility/source/extended/AccessibleBrowseBoxBase.cxx'.

This way, such header files are added to the .pro files
and are thus e.g. shown in Qt Creator's project view
and can be found by using 'a <FILEANME>' in the Locator.

Change-Id: Id74f971b2ffee82203f74a4d444c41166c671920
Reviewed-on: https://gerrit.libreoffice.org/75628
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>