Fix libreoffice version for links in tipoftheday
tipoftheday.hrc has some URL with fixed LibreOffice version (6.0 and 6.2)
Use %PRODUCTVERSION instead and perform the convertion to the correct
version at run-time.

Note, this can have the drawback to point to dead links if a newer
version of the online doc removes a page or changes its path.

'en-US' could also certainly be handled the same way.

Finaly the 'uiResId(aLink.toUtf8().getStr()' to get it converted is
certainly a hack that should be implemented in a better way.

So consider this patch as a PoC :)

Change-Id: Ibdcaedde631f1a243e0ec54cdec6c488c04b8f29
Signed-off-by: Christophe JAILLET <christophe.jaillet@wanadoo.fr>
Reviewed-on: https://gerrit.libreoffice.org/75971
Tested-by: Jenkins
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>