tdf#106000 sc ui: fix missing invalidation when cancelling note creation
Regression from commit 890b943ff5314b6723bf7d5c6f2430b452111c07
(refactor ListBox/ComboBox to use RenderContext, 2015-05-07),
invalidating the area of note anchors worked like this in the past:

- user ended text edit of a new, empty comment
- the Calc shell was updated from shape text edit to the tab view, which
  called ImplListBoxWindow::SelectEntry()
- this called ImplPaint(), which invoked maUserDrawHdl (if it was set)
- and at the end this invoked ScViewFunc::UpdateSelectionArea(), all
  triggered by the font name and font size listboxes

So the update of the cell of the former note depended on the presence of
font name/size listboxes, by accident.

This broke, because ImplListBoxWindow::SelectEntry() now calls
Invalidate(), and Paint() calls ImplDoPaint(), which doesn't invoke
maUserDrawHdl anymore. Fix the problem by invalidating the cell
explicitly, so in case the user customizes the toolbar to not have these
listboxes, the behavior is still correct.

Do this in FuText::StopEditMode(), because there we know the position of
the deleted note and that's where we delete the matching SdrCaptionObj
(via the RemoveLastUndoAction()).

Change-Id: Iaa4b609477ac6c7f23efc575fddac6e39aa6a7b5
Reviewed-on: https://gerrit.libreoffice.org/76120
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>
Tested-by: Jenkins
Reviewed-by: Dennis Francis <dennis.francis@collabora.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>