tdf#107394 sc ui: fix lack of row height update on undo of paste
This is a problem since commit ede5567f8a46fd9971e9f2ad515c5957a655b5bf
(fdo#73606: Avoid excessive and unnecessary heap allocation of array
objects., 2014-01-14), though it seems this uncovered a problem that was
introduced earlier.

Reading ScViewFunc::AdjustBlockHeight(), commit
da82474bb0f32f2f16637cb315becf6e5903f32e (ScUndoPaste to handle multiple
ranges., 2011-09-09) changed ScUndoPaste::DoChange(), so that
AdjustBlockHeight() is invoked in a loop, but always with the same
arguments. In case AdjustBlockHeight() doesn't get a mark data, the
function takes the mark data from the view data, which is empty in the
"undo of paste" case, so no row height gets updated.

Fix the problem by explicitly providing the mark data to
AdjustBlockHeight(), probably that was the intention in the second
commit.

[ Testing this from CppunitTest_sc_copypaste means ScImportExport has to
be dllpublic, but that seems still better than having a view shell in
CppunitTest_sc_ucalc. ]

Change-Id: I140b53588d59d231772152c0e820e5fdedf6894c
Reviewed-on: https://gerrit.libreoffice.org/76265
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>