tdf#118384 Fix paragraph, character, and frame style undo redo
In order for the inherited (parent) style to be restored after create
style undo redo it must be included in the create style undo append.
This patch prevents create style undo append during style make and
provides for it after style dialog return ok when the parent of the new
style is known.

Change-Id: Ie9d7fb7989a5e669db18a1f263632ff208b0e718
Reviewed-on: https://gerrit.libreoffice.org/76279
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>