tdf#126512 make handle of ooxml-shapes usable in odp
The error was that handles in OOXML shapes can be used in LibreOffice
as long as you work in OOXML format, but not anymore when you convert
the file to ODF. Handles in OOXML reference the adjustment values by
name, e.g. 'RefX = adj5'. ODF has no way to save this, so this
information is lost. The patch reconstructs this information from the
shape definition of the preset shape. It gets the preset shape name
'foo' from the shape type 'ooxml-foo'. This means that it only works
with our own naming convention. Still, I think it's an improvement for
our users.

Change-Id: Iebd9f36a5c36356a12c8687e961c7802111cbd85
Reviewed-on: https://gerrit.libreoffice.org/76887
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>