tdf#126723 writerfilter::finishParagraph - me, not previous
In LO 6.2 commit b8c51ee45a5d029c216bbb6ac54ffc1d14ef596e
the existing m_xPreviousParagraph was conveniently used to
apply the changed properties. I never did like that choice,
but despite looking at it, I failed to see that it is set
in an inside loop, which means that it was NOT NECESSARILY
reset to the current paragaph. So I'm happy to have proof
that we should not use m_xPreviousParagraph.

Change-Id: I5c7f1b0f097711d65ae0d0be1f0fbc40c8b96e9d
Reviewed-on: https://gerrit.libreoffice.org/77185
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: László Németh <nemeth@numbertext.org>