sw: fix ~SwIndexReg asserts on ODF export of ooo83072-1.odt
It is now possible that the stupid redline-moving code deletes the first
node in the document, and there are 2 SwPaM that point to it,
Writer::m_pCurrentPam and local pPam in SwWriter::Write().

So i thought it should be quite trivial to just use SwUnoCursors in
these cases, but it required a bit more keyboard bashing than expected.

(regression from c14bd1d2f2191c8527c3bb9b64dd110324bd05a2,
 which i didn't really intend to push and am not sure if it's really
 a good idea but whatever...)

Change-Id: Ia5c18d67af8760664517a4b7ee62ef3e4a417686
Reviewed-on: https://gerrit.libreoffice.org/77225
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>