tdf#126808 vcl: fix crash due to leaking the frame's buffer vdev
This vdev is used to do vcl-level double-buffering, which is off by
default, but e.g. listboxes/comboboxes use it to avoid flicker with
Windows GDI.

The root cause was introduced in commit
56d8601d4f7d4c2c30e2bee8ba7e242d4a54c58f (tdf#92982 vcl: stop
creating/disposing the paint buffer in PaintHelper, 2015-07-30), which
removed the explicit mpBuffer.disposeAndClear() in the PaintHelper dtor,
but forgot to explicitly dispose it elsewhere.

This is now a much more visible problem since commit
ec1a0d19bf21526a026560d15923c77418c72270 (tdf#101978 vcl
combobox/listbox floating window: avoid flicker, 2019-06-17) as the
infrastructure is used for all list/comboboxes.

Fix the problem by explicitly disposing the vdev before deleting the
frame data, so the underlying GDI resource is released, even if we have
a reference cycle.

Change-Id: Ia217bbd78780a214308e497044e73e378ee8ecd6
Reviewed-on: https://gerrit.libreoffice.org/77370
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins