sw: fix assert on export of ooo39250-1.rtf to DOCX
The problem is that a fly anchored at-char inside a text frame is being
exported; in DocxSdrExport::startDMLAnchorInline() the
m_bFlyFrameGraphic is true so an "inline" element is written but then
DocxSdrExport::writeDMLTextFrame() sets the flag to true then false,
so in endDMLAnchorInline() it's false and "anchor" results.

Remove the pointless information hiding between DocxSdrExport and its
own pImpl class so that the previous flag can be restored.

(regression from commit 245b52ee39d03c62998acba87405ead7afc51d71
 which changed things somehow so that the anchor is inside the text
 frame instead of the body as it was before; Word has it also in the
 text frame so we can consider it an improvement)

Change-Id: I2df20b5cc336251f1c20cf81b6d68153c0c6ff46
Reviewed-on: https://gerrit.libreoffice.org/77454
Tested-by: Jenkins
Reviewed-by: Michael Stahl <Michael.Stahl@cib.de>