Gtk3 defer LO dragExit handling to an idle
LO expects to handle either a drop or an exit event. Quoting the
Gtk+ docs: "Likewise, the “drag-leave” signal is also emitted
before the ::drag-drop signal, for instance to allow cleaning up
of a preview item created in the “drag-motion” signal handler."

Since we don't know if there will be a drop-event following the
leave event, we have to defer the "fire_draExit" to an idle, which
will be removed by the drop event, before it can be processed.

Fixes D'n'D between two Writer windows. For whatever reason it
works correct in Calc, which probably needs further investigation.

Regression from commit e00c33a68f76 ("Gtk3 inform LO of drag
leave").

Change-Id: Ia59dcc8b20efd442ac5efc895602e5cc432cc379
Reviewed-on: https://gerrit.libreoffice.org/77753
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>