More hacks for quit requests from an OLE Automation client
Actually I am now not so sure whether the TerminationVetoer thing is
needed or not. Will have to experiment later with the customer use
case what happens if I remove all that. Maybe adding it was a mistake,
as misinterpretation of what was going on.

Change-Id: I252e1233cae1622099bc3310814132dae58b2aed
(cherry picked from commit 9c50273a9f3a15d1ff7ea4880638ca833570e0bd)
Reviewed-on: https://gerrit.libreoffice.org/77948
Tested-by: Jenkins
Reviewed-by: Tor Lillqvist <tml@collabora.com>
Tested-by: Tor Lillqvist <tml@collabora.com>