tdf#93998 toolkit: restore support for setting dialog background from shape
Regression from commit c2ac98d45bc4cce859b3d47542d5a652ec71de2d (svx:
removing GraphicURL and OWN_ATTR_GRAFURL, fix writerfilter, 2018-03-08),
the problem was that an in-document macro tried to assign a bitmap shape
to a dialog background, which broke during the image handling rework.

Note that in this case the actual type of the "URL" is not interesting,
we can just return an XGraphic and then take it on the other side as
well, without re-introducing the intentionally removed graphic URLs
which point to memory addresses.

This also made it necessary to extend
UnoDialogControl::ImplModelPropertiesChanged(), so that in case a
graphic is assigned to the dialog model, then the dialog model -> dialog
sync code doesn't just copy over the empty image URL string. With this,
finally clicking on the button of the bugdoc makes the dialog show up
with the correct background.

Change-Id: Id78269643289efb435b96a6a0b9f8a93fa49ec04
Reviewed-on: https://gerrit.libreoffice.org/78153
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins