Simplify code removing the last segment from a URL
This code had been introduced with d1d6f5c444d3313badb3c3ad6cf872908026e680
"tdf#86087 Save relative links in DOCX", but it is not clear to me what part
of <https://bugs.documentfoundation.org/show_bug.cgi?id=86087> "FILESAVE
FILEOPEN VIEWING: Can't open or save relative links in docx" it is meant to
address.  None of the tests in `make check` trigger the

>         else
>         {
>             // DOC
>             WW8Export* pWW8Export = dynamic_cast<WW8Export*>(&GetExport());
>             if ( pWW8Export )
>             {
>                 SwWW8Writer& rWriter = pWW8Export->GetWriter();
>                 sExportedDocumentURL = rWriter.GetMedia()->GetURLObject().GetPath();
>             }
>         }

block.

But what the code presumably wants to do is to create, for the DOC case, an
anAbsoluteParent that is rWriter.GetMedia()->GetURLObject() with the final URL
path segment removed:  INetURLObject::GetPath returns the "dirname" part of a
file URL's path (i.e., everything but the last path segment), and with a
trailing slash removed, in the form of a filesystem pathname.  That means that

>     INetURLObject anAbsoluteParent( sExportedDocumentURL );

creates an invalid anAbsoluteParent (because the input is not a URL), so

>     if ( anAbsoluteParent.GetURLPath().isEmpty() )

is true (but why not use a more obvious check of anAbsoluteParent.HasError(),
or should this code also be relevant for the DOCX case?), so the following block

>     {
>         // DOC filter returns system path (without file:///)
>         anAbsoluteParent.setFSysPath( sExportedDocumentURL, FSysStyle::Detect );
>         anAbsoluteParent.setFinalSlash();
>     }

recreated the URL from the filesystem pathname and added back the trailing
slash.

Change-Id: I9346a23680a65b6fd1ba0fa33f4565df50bd9e51
Reviewed-on: https://gerrit.libreoffice.org/78210
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>