tdf#127211 start center: fix background of the main area with no recent docs
The background was white instead of gray, but the text color remained
white, making the text unreadable. This was a problem since commit
3516b727dfe569f2d1bfdcd6a216f5bcf9985ab1 (tdf#91843 sd sidebar: fix
unexpected borders around the slide layout previews, 2019-08-27).

Given that the background fix was intended to help with the slide
previews in impress, move the change affecting the background handling
down to ValueSet and revert the VCL changes.

This preserves the impress slide preview fix, but restores the old good
behavior for all non-ValueSet widgets, including the start center.

Change-Id: I7efcd5b9c12e5df8e10261c992bf7a774960fa11
Reviewed-on: https://gerrit.libreoffice.org/78280
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins