[API CHANGE] remove MacroMigrationWizard
https://wiki.openoffice.org/wiki/Macros_in_Database_Documents

Prior to OpenOffice 3 (2008) base didn't support macros, but
subdocuments might have had macros. since OOo3 base supported macros
and subdocuments not, and a migration wizard is available to update
pre OOo3 documents to the new scheme.

Here I presume whatever is going to get migrated has been migrated
at this point and drop the migration wizard

This undoes the addition of the idl for
sdb::application::MacroMigrationWizard of

commit 83dd9931ae7f0ae9fcc2f7aa0f7b80f335d18299
Date:   Mon Mar 11 17:51:56 2013 +0200

so flag as an api change, though not to anything published

Change-Id: I37271752234dda4e7e8f033e0136825fd356439b
Reviewed-on: https://gerrit.libreoffice.org/78326
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>